Optimized the splitting of hunks
Previously each single subhunk required a call to splithunk.py.
Unfortunately, due to the startup overhead of python, this slowed the
entire process down significantly.

To improve this, splithunk.py has been modified to be able to handle
the splitting of a full patch into all available subhunks in a single
call. This means that a bunch of filterdiff + splithunk.py calls have
been combined into a single splithunk.py call.

This seems to result in a speed improvement of 4-5x the previous speed.